#1 Move User to Models Folder, fix PHP docs
